OBG-509. Cleanup structure
