+++ opba-db/src/main/java/de/adorsys/opba/db/domain/entity/psu/PsuPrivate.java
@@ -7,12 +7,13 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Basic;
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.Id;
 import javax.persistence.Lob;
-import java.sql.Blob;
+import javax.persistence.ManyToOne;
 import java.util.UUID;
 
 @Getter
@@ -26,8 +27,11 @@ public class PsuPrivate {
     @Id
     private UUID id;
 
+    @ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY, optional = false)
+    private Psu psu;
+
     @Lob
     @Basic(fetch = FetchType.LAZY)
     @Column(nullable = false)
-    private Blob data;
+    private byte[] data;
 }
